# contracts/openapi/v2/MCP-001-grader-api.yaml
...
openapi: 3.0.3
...
info:
  title: Smackdab MCP API Grader Service
  description: |
    # MCP API Grader Service
    
    Enterprise-grade API contract validation and grading service that ensures OpenAPI specifications meet Smackdab's rigorous quality standards. This service provides comprehensive API contract analysis, scoring, and improvement recommendations.
    
    ## Key Features
    - **Comprehensive API Grading**: Validates OpenAPI 3.0.3 specifications against 100+ quality checkpoints
    - **Real-time Validation**: Instant feedback on API contract compliance and best practices
    - **Historical Tracking**: Maintains grading history for continuous improvement tracking
    - **Automated Fixes**: Suggests and applies corrections to common API design issues
    - **Multi-tenant Support**: Full tenant isolation with organization and branch context
    
    ## Business Rules Implemented
    - **RULE-GRD-001**: All APIs must achieve minimum 80% compliance for production deployment
    - **RULE-GRD-002**: Critical security violations trigger automatic grade failure
    - **RULE-GRD-003**: Multi-tenant headers mandatory on all endpoints
    - **RULE-GRD-004**: Rate limiting enforced per tenant with tiered quotas
    - **RULE-GRD-005**: API contracts must follow HATEOAS Level 3 REST principles
    
    ## Technical Architecture
    - **Database**: PostgreSQL with Citus distribution by organization_id
    - **Caching**: Dragonfly/Valkey for grading result caching
...
    
...
    
...
    
...

...
servers:
...
security:
  - OAuth2: []
  - BearerAuth: []

# ============================================================================
# PATHS
# ============================================================================
paths:
...
  /api/v2/grader/checkpoints:
    get:
      summary: List all grading checkpoints
      description: Returns the complete list of grading checkpoints used for API validation with their weights and categories
      operationId: listCheckpoints
      tags:
        - System
      security:
        - OAuth2: ["grader.read"]
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
        - $ref: '#/components/parameters/BranchHeader'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Checkpoints retrieved successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
            ETag:
              schema:
                type: string
                example: '"abc123def456"'
            Cache-Control:
              schema:
                type: string
                example: "public, max-age=86400"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckpointsResponse'
              examples:
                standard:
                  summary: Standard checkpoints list
                  value:
                    success: true
                    data:
                      - id: "OAS-VERSION"
                        category: "structure"
                        description: "OpenAPI version must be 3.0.3"
                        weight: 10
                        auto_fail: true
                      - id: "MULTI-TENANT"
                        category: "security"
                        description: "Multi-tenant headers required"
                        weight: 15
                        auto_fail: true
                      - id: "HATEOAS"
                        category: "rest"
                        description: "HATEOAS Level 3 compliance"
                        weight: 20
                        auto_fail: false
                    meta:
                      request_id: "req_789ghi"
                      timestamp: "2025-01-10T12:00:00Z"
                      total_checkpoints: 120
                      total_weight: 100
        '304':
          $ref: '#/components/responses/NotModified'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v2/grader/grade-inline:
    post:
      summary: Grade inline YAML content
      description: Grades OpenAPI specification content provided directly in the request body
      operationId: gradeInline
      tags:
        - Grading
      security:
        - OAuth2: ["grader.write"]
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
        - $ref: '#/components/parameters/BranchHeader'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              type: string
              description: Raw YAML content of the OpenAPI specification
            example: |
              openapi: 3.0.3
              info:
                title: Sample API
                version: 1.0.0
              paths:
                /api/v2/resources:
                  get:
                    summary: List resources
                    responses:
                      '200':
                        description: Success
      responses:
        '200':
          description: Grading completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
...
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.smackdab.com/oauth/authorize
          tokenUrl: https://auth.smackdab.com/oauth/token
          scopes:
            grader.read: Read grading results and checkpoints
            grader.write: Perform grading operations
            grader.admin: Administrative operations
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing:
        - org_id: Organization UUID
        - branch_id: Branch UUID
        - user_id: User UUID
        - permissions: Array of permission strings

  # ============================================================================
  # PARAMETERS
  # ============================================================================
  parameters:
    OrganizationHeader:
      name: X-Organization-ID
      in: header
      required: true
      description: Organization identifier for multi-tenant isolation
      schema:
        type: string
        format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
    BranchHeader:
      name: X-Branch-ID
      in: header
      required: true
      description: Branch identifier for multi-tenant isolation
      schema:
        type: string
        format: uuid
      example: "456e7890-e89b-12d3-a456-426614174111"
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Unique request identifier for tracing
      schema:
        type: string
        format: uuid
      example: "789e0123-e89b-12d3-a456-426614174222"
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: false
      description: Idempotency key for safe retries
      schema:
        type: string
        minLength: 16
        maxLength: 64
      example: "unique-key-123abc"
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: ETag for conditional requests
      schema:
        type: string
      example: '"abc123def456"'
...
  headers:
    RateLimitLimit:
      description: Rate limit ceiling for current tier
      schema:
        type: integer
      example: 100
    RateLimitRemaining:
      description: Requests remaining in current window
      schema:
        type: integer
      example: 95
    RateLimitReset:
      description: UTC timestamp when rate limit resets
      schema:
        type: integer
      example: 1704934800

  # ============================================================================
  # SCHEMAS
  # ============================================================================
  schemas:
...
    ResponseEnvelope:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        data:
          type: object
          description: Response payload
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        _links:
          $ref: '#/components/schemas/HATEOASLinks'
    ResponseMeta:
      type: object
      required:
        - request_id
        - timestamp
      properties:
        request_id:
          type: string
          format: uuid
          description: Unique request identifier
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        processing_time_ms:
          type: integer
          description: Processing time in milliseconds
        warnings:
          type: array
          items:
            type: string
          description: Non-critical warnings
    HATEOASLinks:
      type: object
      properties:
        self:
          type: object
          properties:
            href:
              type: string
              format: uri
        collection:
          type: object
          properties:
            href:
              type: string
              format: uri
        next:
          type: object
          properties:
            href:
              type: string
              format: uri
        prev:
          type: object
          properties:
            href:
              type: string
              format: uri

    # Version Response
...
    CheckpointsResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Checkpoint'
            meta:
              allOf:
                - $ref: '#/components/schemas/ResponseMeta'
                - type: object
                  properties:
                    total_checkpoints:
                      type: integer
                    total_weight:
                      type: integer
    Checkpoint:
      type: object
      required:
        - id
        - category
        - description
        - weight
        - auto_fail
      properties:
        id:
          type: string
          example: "OAS-VERSION"
        category:
          type: string
          enum: [structure, security, rest, documentation, performance, webhooks]
        description:
          type: string
        weight:
          type: integer
          minimum: 0
          maximum: 100
        auto_fail:
          type: boolean
          description: Whether violation triggers automatic failure

    # Grade Request
...
    GradeResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
        - type: object
          properties:
            data:
              type: object
              required:
                - grade
                - findings
                - metadata
              properties:
                grade:
                  $ref: '#/components/schemas/Grade'
                findings:
                  type: array
                  items:
                    $ref: '#/components/schemas/Finding'
                checkpoints:
                  type: array
                  items:
                    $ref: '#/components/schemas/CheckpointScore'
                metadata:
                  $ref: '#/components/schemas/GradingMetadata'
    Grade:
      type: object
      required:
        - total
        - letter
        - compliance_pct
        - auto_fail_triggered
        - critical_issues
      properties:
        total:
          type: number
          minimum: 0
          maximum: 100
        letter:
          type: string
          enum: [A+, A, A-, B+, B, B-, C, D, F]
        compliance_pct:
          type: number
          minimum: 0
          maximum: 1
        auto_fail_triggered:
          type: boolean
        critical_issues:
          type: integer
          minimum: 0
        auto_fail_reasons:
          type: array
          items:
            type: string
        per_category:
          type: object
          additionalProperties:
            type: object
            properties:
              earned:
                type: number
              max:
                type: number
              percentage:
                type: number
        coverage_based:
          type: boolean
          description: Whether coverage-based scoring was used
        excellence:
          type: object
          description: Excellence metrics for high-scoring APIs
    Finding:
      type: object
      required:
        - rule_id
        - message
        - severity
        - json_path
      properties:
        rule_id:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [error, warning, info]
        json_path:
          type: string
        category:
          type: string
        line:
          type: integer
    CheckpointScore:
      type: object
      required:
        - checkpoint_id
        - category
        - max_points
        - scored_points
      properties:
        checkpoint_id:
          type: string
        category:
          type: string
        max_points:
          type: integer
        scored_points:
          type: integer
    GradingMetadata:
      type: object
      required:
        - spec_hash
        - template_hash
        - ruleset_hash
        - template_version
        - scoring_engine
      properties:
        spec_hash:
          type: string
          pattern: '^sha256:[a-f0-9]{64}$'
        template_hash:
          type: string
          pattern: '^sha256:[a-f0-9]{64}$'
        ruleset_hash:
          type: string
          pattern: '^sha256:[a-f0-9]{64}$'
        template_version:
          type: string
        scoring_engine:
          type: string
          enum: [coverage-based, legacy]
        instance_id:
          type: string
        instance_start_time:
          type: string
          format: date-time
        graded_at:
          type: string
          format: date-time

    # Grade and Record Request
...
    ProblemDetail:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          format: uri
          example: "https://api.smackdab.com/problems/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "The request contains invalid field values"
        instance:
          type: string
          format: uri
          example: "/api/v2/grader/grade"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "/data/content"
              code:
                type: string
                example: "REQUIRED"
              message:
                type: string
                example: "Content field is required"

  # ============================================================================
  # RESPONSES
  # ============================================================================
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: "https://api.smackdab.com/problems/bad-request"
            title: "Bad Request"
            status: 400
            detail: "Invalid request parameters"
            instance: "/api/v2/grader/grade"
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: "https://api.smackdab.com/problems/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "Invalid or missing authentication token"
            instance: "/api/v2/grader/grade"
    Forbidden:
      description: Insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: "https://api.smackdab.com/problems/forbidden"
            title: "Forbidden"
            status: 403
            detail: "You don't have permission to access this resource"
            instance: "/api/v2/grader/grade"
...
    UnsupportedMediaType:
      description: Unsupported media type
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: "https://api.smackdab.com/problems/unsupported-media-type"
            title: "Unsupported Media Type"
            status: 415
            detail: "Content-Type must be application/json or application/x-yaml"
            instance: "/api/v2/grader/grade"
    UnprocessableEntity:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: "https://api.smackdab.com/problems/validation-error"
            title: "Validation Error"
            status: 422
            detail: "The OpenAPI specification is not valid YAML"
            instance: "/api/v2/grader/grade"
            errors:
              - field: "/data/content"
                code: "INVALID_YAML"
                message: "Content is not valid YAML syntax"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: "https://api.smackdab.com/problems/rate-limit-exceeded"
            title: "Too Many Requests"
            status: 429
            detail: "Rate limit exceeded. Please retry after 60 seconds"
            instance: "/api/v2/grader/grade"
    InternalServerError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: "https://api.smackdab.com/problems/internal-error"
            title: "Internal Server Error"
            status: 500
            detail: "An unexpected error occurred"
            instance: "/api/v2/grader/grade"
    ServiceUnavailable:
      description: Service temporarily unavailable
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until service is available
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: "https://api.smackdab.com/problems/service-unavailable"
            title: "Service Unavailable"
            status: 503
            detail: "Service is temporarily unavailable for maintenance"
            instance: "/api/v2/grader/grade"
    NotModified:
      description: Resource not modified (cached response valid)